     \name{wheat}
     \docType{data}
     \alias{wheat}
     \title{Genetic and Phenotype Data of a Wheat Breeding Population}
     \description{
       This data set gives the genetic marker information (DArT markers), phenotype, and the kinship matrix of a wheat breeding population.  The list wheat contains three elelments, 'y', 'marker' and 'A'.  The vector 'y' contains the phenotypic values for 282 wheat accessions, 'A' is a relationship (kinship) matrix of all the wheat accessions, and 'marker' is a 282 by 300 data frame for the genotype on 300 DArT markers for the wheat accessions.
     }
     \usage{data(wheat)}
     \format{ A list of three elements.    }
     \source{Nebraska Wheat Breeding}
     \references{
      Wang, D., \emph{et al.} (2012) Prediction of genetic values of quantitative traits with epistatic effects in plant breeding populations. \emph{Heredity}, 109: 313-319.
     }
     \keyword{datasets}
